+++ app/src/main/java/com/money/manager/ex/investment/PortfolioCursorAdapter.java
@@ -72,7 +72,7 @@ public class PortfolioCursorAdapter
         PortfolioItemDataViewHolder holder = (PortfolioItemDataViewHolder) view.getTag();
 
         // header index
-        long accountId = cursor.getInt(cursor.getColumnIndex(StockFields.HELDAT));
+        long accountId = cursor.getLong(cursor.getColumnIndex(StockFields.HELDAT));
         if (!mHeadersAccountIndex.containsKey(accountId)) {
             mHeadersAccountIndex.put(accountId, cursor.getPosition());
         }
